+++ app/src/main/java/rakki/inventory/basic/authendication/UserPreference.kt
@@ -5,13 +5,13 @@ import android.content.SharedPreferences
 
 class UserPreference(context: Context) {
     private val sharePreference: SharedPreferences = context.getSharedPreferences("user", Context.MODE_PRIVATE)
-    fun setUserInfoToFile(userId: Int) {
+    fun setUserInfoToFile(userId: Long) {
         val edit = sharePreference.edit()
-        edit.putInt("userId", userId)
+        edit.putLong("userId", userId)
         edit.apply()
     }
 
-    fun getLoggedUser(): Int {
-        return sharePreference.getInt("userId", -1)
+    fun getLoggedUser(): Long {
+        return sharePreference.getLong("userId", -1)
     }
 }
